lsm: Preserve full ioctl commands in audit records
Ioctl commands are unsigned int values, with their encoded direction and
size in the upper 16 bits.  However, struct lsm_ioctlop_audit keeps only
a u16, and the common audit formatter uses %hx.  This logs 0xc00ffeee as
0xfeee and makes distinct commands with matching low 16 bits
indistinguishable.

Change cmd to unsigned int and use %x.  Pass the full command through
ioctl_has_perm() as well.  Its driver and xperm values remain u8 values
derived from the same low 16 bits, so SELinux enforcement does not
change.  Commands that fit in 16 bits keep the same audit text.

Cc: Jeff Vander Stoep <jeffv@google.com>
Cc: Paul Moore <paul@paul-moore.com>
Fixes: 671a2781ff01 ("security: add ioctl specific auditing to lsm_audit")

+++ include/linux/lsm_audit.h
@@ -44,7 +44,7 @@ struct lsm_network_audit {
 
 struct lsm_ioctlop_audit {
 	struct path path;
-	u16 cmd;
+	unsigned int cmd;
 };
 
 struct lsm_ibpkey_audit {
